requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.5.0
requests==2.31.0
celery==5.3.4
django-celery-beat==2.5.0
